1-install: use the cpio from android-x86 system
There are two busybox in android-x86. One is static linked and the other
is linked with bionic. The prior is older version and only used in boot
stage and installer, the later is newer and can be used after booted
into android-x86.

It seems the prior busybox cpio is buggy to copy files with set-user-ID bit.
Using /system/xbin/cpio solves the issue.